+++ packages/velo-external-db-core/lib/converters/data_utils.js
@@ -1,0 +1,17 @@
+const { isDate } = require('velo-external-db-commons')
+const crypto = require('crypto')
+
+const asWixData = e => generateIdsIfNeeded(packDates(e))
+
+const generateIdsIfNeeded = item => {
+    if ('_id' in item)
+        return item
+    const sha = crypto.createHash('sha1')
+    const fieldsConcat = Object.values(item).join('')
+    return { ...item, _id: sha.update(fieldsConcat).digest('base64') }
+}
+
+const packDates = item => Object.entries(item)
+                                .reduce((o, [k, v]) => ({ ...o, [k]: isDate(v) ? { $date: v.toISOString() } : v }), {})
+
+module.exports = { asWixData, generateIdsIfNeeded }
